+++ Changelog.md
@@ -34,6 +34,7 @@
    * Mercenary licence now obtained from a mission
    * Llama voyager no longer has a super efficient jumpdrive
    * Nerfed range on torpedoes and other heavy launcher weapons
+   * Boarding bonus affects how much fleet capacity you need to capture ships
 
  ### Content